feat: share jwt guard via auth package
