+++ services/api-gateway/src/routes/admin.data.ts
@@ -1,13 +1,19 @@
 import type { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
-import { authenticateRequest, type Role } from "../lib/auth.js";
+import { authenticateRequest, type Role } from "@apgms/auth";
 
 // minimal in-memory placeholder dataset
 const DATA: { id: string; value: string; orgId: string }[] = [];
 
 export async function registerAdminDataRoutes(app: FastifyInstance) {
   const gate = (roles: readonly Role[] = []) =>
-    async (req: FastifyRequest, reply: FastifyReply) =>
-      authenticateRequest(app, req, reply, roles);
+    async (req: FastifyRequest, reply: FastifyReply) => {
+      const principal = await authenticateRequest(app, req, reply, roles);
+      if (!principal) {
+        return;
+      }
+
+      (req as any).user = principal;
+    };
 
   app.get("/admin/data", { preHandler: gate([]) }, async (req, reply) => {
     const orgId = (req as any).user?.orgId as string | undefined;
